:recycle: Extract auth.go from pipeline_handler.go
