+++ src/api/auth.go
@@ -1,0 +1,37 @@
+package api
+
+import (
+	"net/http"
+	"regexp"
+
+	"models"
+
+	"github.com/labstack/echo"
+	"golang.org/x/net/context"
+)
+
+const (
+	AUTH_HEADER = "Authorization"
+)
+
+func (h *PipelineHandler) withAuth(impl func(c echo.Context) error) func(c echo.Context) error {
+	return func(c echo.Context) error {
+		ctx := c.Get("aecontext").(context.Context)
+		req := c.Request()
+		raw := req.Header.Get(AUTH_HEADER)
+		if raw == "" {
+			return c.JSON(http.StatusUnauthorized, map[string]string{"message": "Unauthorized"})
+		}
+		re := regexp.MustCompile(`\ABearer\s+`)
+		token := re.ReplaceAllString(raw, "")
+		if token == "" {
+			return c.JSON(http.StatusUnauthorized, map[string]string{"message": "Unauthorized"})
+		}
+		org := c.Get("organization").(*models.Organization)
+		_, err := org.AuthAccessor().FindWithToken(ctx, token)
+		if err != nil {
+			return c.JSON(http.StatusUnauthorized, map[string]string{"message": "Invalid token"})
+		}
+		return impl(c)
+	}
+}
